refactor(telephony): split routes/telephony.py per provider
Each provider now owns its webhooks, status callbacks, and answer-URL
endpoints in providers/<name>/routes.py and registers them via the
ProviderSpec.router field added in the previous commit. The main
routes/telephony.py keeps shared concerns: /initiate-call, the
WebSocket entry, generic /inbound/{workflow_id}, transfer endpoints,
and the registry-driven include_router loop that mounts each provider.

Routes moved (per provider):
  twilio:    /twiml, /twilio/status-callback/{workflow_run_id}
  vonage:    /ncco, /vonage/events/{workflow_run_id}
  telnyx:    /telnyx/events/{workflow_run_id}
  vobiz:     /vobiz-xml, /vobiz/{hangup,ring}-callback/{workflow_run_id},
             /vobiz/hangup-callback/workflow/{workflow_id}
  plivo:     /plivo-xml, /plivo/{hangup,ring}-callback/{workflow_run_id}
             (plus the _handle_plivo_status_callback helper)
  cloudonix: /cloudonix/status-callback/{workflow_run_id}, /cloudonix/cdr

Net: routes/telephony.py shrinks from 1875 -> ~982 lines.

The provider package previously triggered registration at module import,
which broke once provider routes started importing back from factory.
factory.py now lazy-loads the providers package on first call instead,
so module-level imports from provider routes don't cycle.

Webhook URL paths are unchanged — anything configured at provider
consoles continues to work.
