+++ api/routes/telephony.py
@@ -455,45 +455,6 @@ async def _validate_organization_provider_config(
         return TelephonyError.ACCOUNT_VALIDATION_FAILED
 
 
-@router.post("/twiml", include_in_schema=False)
-async def handle_twiml_webhook(
-    workflow_id: int, user_id: int, workflow_run_id: int, organization_id: int
-):
-    """
-    Handle initial webhook from telephony provider.
-    Returns provider-specific response (e.g., TwiML for Twilio).
-    """
-
-    provider = await get_telephony_provider(organization_id)
-
-    response_content = await provider.get_webhook_response(
-        workflow_id, user_id, workflow_run_id
-    )
-
-    return HTMLResponse(content=response_content, media_type="application/xml")
-
-
-@router.get("/ncco", include_in_schema=False)
-async def handle_ncco_webhook(
-    workflow_id: int,
-    user_id: int,
-    workflow_run_id: int,
-    organization_id: Optional[int] = None,
-):
-    """Handle NCCO (Nexmo Call Control Objects) webhook for Vonage.
-
-    Returns JSON response instead of XML like TwiML.
-    """
-
-    provider = await get_telephony_provider(organization_id or user_id)
-
-    response_content = await provider.get_webhook_response(
-        workflow_id, user_id, workflow_run_id
-    )
-
-    return json.loads(response_content)
-
-
 @router.websocket("/ws/ari")
 async def websocket_ari_endpoint(websocket: WebSocket):
     """WebSocket endpoint for ARI chan_websocket external media.
@@ -620,807 +581,6 @@ async def _handle_telephony_websocket(
         except RuntimeError:
             # WebSocket already closed, ignore
             pass
-
-
-@router.post("/twilio/status-callback/{workflow_run_id}")
-async def handle_twilio_status_callback(
-    workflow_run_id: int,
-    request: Request,
-    x_webhook_signature: Optional[str] = Header(None),
-):
-    """Handle Twilio-specific status callbacks."""
-    set_current_run_id(workflow_run_id)
-
-    # Parse form data
-    form_data = await request.form()
-    callback_data = dict(form_data)
-
-    logger.info(
-        f"[run {workflow_run_id}] Received status callback: {json.dumps(callback_data)}"
-    )
-
-    # Get workflow run to find organization
-    workflow_run = await db_client.get_workflow_run_by_id(workflow_run_id)
-    if not workflow_run:
-        logger.warning(f"Workflow run {workflow_run_id} not found for status callback")
-        return {"status": "ignored", "reason": "workflow_run_not_found"}
-
-    # Get workflow and provider
-    workflow = await db_client.get_workflow_by_id(workflow_run.workflow_id)
-    if not workflow:
-        logger.warning(f"Workflow {workflow_run.workflow_id} not found")
-        return {"status": "ignored", "reason": "workflow_not_found"}
-
-    provider = await get_telephony_provider(workflow.organization_id)
-
-    if x_webhook_signature:
-        backend_endpoint, _ = await get_backend_endpoints()
-        full_url = f"{backend_endpoint}/api/v1/telephony/twilio/status-callback/{workflow_run_id}"
-
-        is_valid = await provider.verify_webhook_signature(
-            full_url, callback_data, x_webhook_signature
-        )
-
-        if not is_valid:
-            logger.warning(
-                f"Invalid webhook signature for workflow run {workflow_run_id}"
-            )
-            return {"status": "error", "reason": "invalid_signature"}
-
-    # Parse the callback data into generic format
-    parsed_data = provider.parse_status_callback(callback_data)
-
-    # Create StatusCallbackRequest from parsed data
-    status_update = StatusCallbackRequest(
-        call_id=parsed_data["call_id"],
-        status=parsed_data["status"],
-        from_number=parsed_data.get("from_number"),
-        to_number=parsed_data.get("to_number"),
-        direction=parsed_data.get("direction"),
-        duration=parsed_data.get("duration"),
-        extra=parsed_data.get("extra", {}),
-    )
-
-    # Process the status update
-    await _process_status_update(workflow_run_id, status_update)
-
-    return {"status": "success"}
-
-
-@router.post("/telnyx/events/{workflow_run_id}")
-async def handle_telnyx_events(
-    request: Request,
-    workflow_run_id: int,
-):
-    """Handle Telnyx Call Control webhook events.
-
-    Telnyx sends all call lifecycle events (call.initiated, call.answered,
-    call.hangup, streaming.started, streaming.stopped) as JSON POST requests.
-    """
-    set_current_run_id(workflow_run_id)
-
-    event_data = await request.json()
-    logger.info(
-        f"[run {workflow_run_id}] Received Telnyx event: {json.dumps(event_data)}"
-    )
-
-    # Extract event type from Telnyx envelope
-    data = event_data.get("data", {})
-    event_type = data.get("event_type", "")
-
-    # Skip streaming events — they're informational only
-    if event_type in ("streaming.started", "streaming.stopped"):
-        logger.debug(f"[run {workflow_run_id}] Telnyx streaming event: {event_type}")
-        return {"status": "success"}
-
-    # Get workflow run and provider
-    workflow_run = await db_client.get_workflow_run_by_id(workflow_run_id)
-    if not workflow_run:
-        logger.warning(f"Workflow run {workflow_run_id} not found for Telnyx event")
-        return {"status": "ignored", "reason": "workflow_run_not_found"}
-
-    workflow = await db_client.get_workflow_by_id(workflow_run.workflow_id)
-    if not workflow:
-        logger.warning(f"Workflow {workflow_run.workflow_id} not found")
-        return {"status": "ignored", "reason": "workflow_not_found"}
-
-    provider = await get_telephony_provider(workflow.organization_id)
-
-    # Parse the callback data into generic format
-    parsed_data = provider.parse_status_callback(event_data)
-
-    status_update = StatusCallbackRequest(
-        call_id=parsed_data["call_id"],
-        status=parsed_data["status"],
-        from_number=parsed_data.get("from_number"),
-        to_number=parsed_data.get("to_number"),
-        direction=parsed_data.get("direction"),
-        duration=parsed_data.get("duration"),
-        extra=parsed_data.get("extra", {}),
-    )
-
-    await _process_status_update(workflow_run_id, status_update)
-
-    return {"status": "success"}
-
-
-@router.post("/vonage/events/{workflow_run_id}")
-async def handle_vonage_events(
-    request: Request,
-    workflow_run_id: int,
-):
-    """Handle Vonage-specific event webhooks.
-
-    Vonage sends all call events to a single endpoint.
-    Events include: started, ringing, answered, complete, failed, etc.
-    """
-    set_current_run_id(workflow_run_id)
-    # Parse the event data
-    event_data = await request.json()
-    logger.info(f"[run {workflow_run_id}] Received Vonage event: {event_data}")
-
-    # Get workflow run for processing
-    workflow_run = await db_client.get_workflow_run_by_id(workflow_run_id)
-    if not workflow_run:
-        logger.error(f"[run {workflow_run_id}] Workflow run not found")
-        return {"status": "error", "message": "Workflow run not found"}
-
-    # For a completed call that includes cost info, capture it immediately
-    if event_data.get("status") == "completed":
-        # Vonage sometimes includes price info in the webhook
-        if "price" in event_data or "rate" in event_data:
-            try:
-                if workflow_run.cost_info:
-                    # Store immediate cost info if available
-                    cost_info = workflow_run.cost_info.copy()
-                    if "price" in event_data:
-                        cost_info["vonage_webhook_price"] = float(event_data["price"])
-                    if "rate" in event_data:
-                        cost_info["vonage_webhook_rate"] = float(event_data["rate"])
-                    if "duration" in event_data:
-                        cost_info["vonage_webhook_duration"] = int(
-                            event_data["duration"]
-                        )
-
-                    await db_client.update_workflow_run(
-                        run_id=workflow_run_id, cost_info=cost_info
-                    )
-                    logger.info(
-                        f"[run {workflow_run_id}] Captured Vonage cost info from webhook"
-                    )
-            except Exception as e:
-                logger.error(
-                    f"[run {workflow_run_id}] Failed to capture Vonage cost from webhook: {e}"
-                )
-
-    # Get workflow and provider
-    workflow = await db_client.get_workflow_by_id(workflow_run.workflow_id)
-    if not workflow:
-        logger.error(f"[run {workflow_run_id}] Workflow not found")
-        return {"status": "error", "message": "Workflow not found"}
-
-    provider = await get_telephony_provider(workflow.organization_id)
-
-    # Parse the event data into generic format
-    parsed_data = provider.parse_status_callback(event_data)
-
-    # Create StatusCallbackRequest from parsed data
-    status_update = StatusCallbackRequest(
-        call_id=parsed_data["call_id"],
-        status=parsed_data["status"],
-        from_number=parsed_data.get("from_number"),
-        to_number=parsed_data.get("to_number"),
-        direction=parsed_data.get("direction"),
-        duration=parsed_data.get("duration"),
-        extra=parsed_data.get("extra", {}),
-    )
-
-    # Process the status update
-    await _process_status_update(workflow_run_id, status_update)
-
-    # Return 204 No Content as expected by Vonage
-    return {"status": "ok"}
-
-
-@router.post("/vobiz-xml", include_in_schema=False)
-async def handle_vobiz_xml_webhook(
-    workflow_id: int, user_id: int, workflow_run_id: int, organization_id: int
-):
-    """
-    Handle initial webhook from Vobiz when call is answered.
-    Returns Vobiz XML response with Stream element.
-
-    Vobiz uses Plivo-compatible XML format similar to Twilio's TwiML.
-    """
-    set_current_run_id(workflow_run_id)
-    logger.info(
-        f"[run {workflow_run_id}] Vobiz XML webhook called - "
-        f"workflow_id={workflow_id}, user_id={user_id}, org_id={organization_id}"
-    )
-
-    provider = await get_telephony_provider(organization_id)
-
-    logger.debug(f"[run {workflow_run_id}] Using provider: {provider.PROVIDER_NAME}")
-
-    response_content = await provider.get_webhook_response(
-        workflow_id, user_id, workflow_run_id
-    )
-
-    logger.debug(
-        f"[run {workflow_run_id}] Vobiz XML response generated:\n{response_content}"
-    )
-
-    return HTMLResponse(content=response_content, media_type="application/xml")
-
-
-@router.post("/plivo-xml", include_in_schema=False)
-async def handle_plivo_xml_webhook(
-    workflow_id: int,
-    user_id: int,
-    workflow_run_id: int,
-    organization_id: int,
-    request: Request,
-    x_plivo_signature_v3: Optional[str] = Header(None),
-    x_plivo_signature_ma_v3: Optional[str] = Header(None),
-    x_plivo_signature_v3_nonce: Optional[str] = Header(None),
-):
-    """
-    Handle initial webhook from Plivo when an outbound call is answered.
-    Returns Plivo XML response with Stream element.
-    """
-    set_current_run_id(workflow_run_id)
-    provider = await get_telephony_provider(organization_id)
-
-    form_data = await request.form()
-    callback_data = dict(form_data)
-
-    signature = x_plivo_signature_v3 or x_plivo_signature_ma_v3
-    if signature:
-        backend_endpoint, _ = await get_backend_endpoints()
-        full_url = (
-            f"{backend_endpoint}/api/v1/telephony/plivo-xml"
-            f"?workflow_id={workflow_id}"
-            f"&user_id={user_id}"
-            f"&workflow_run_id={workflow_run_id}"
-            f"&organization_id={organization_id}"
-        )
-        is_valid = await provider.verify_inbound_signature(
-            full_url, callback_data, signature, x_plivo_signature_v3_nonce
-        )
-        if not is_valid:
-            logger.warning(
-                f"[run {workflow_run_id}] Invalid Plivo signature on answer webhook"
-            )
-            return provider.generate_error_response(
-                "invalid_signature", "Invalid webhook signature."
-            )
-
-    call_id = callback_data.get("CallUUID") or callback_data.get("RequestUUID")
-    if call_id:
-        workflow_run = await db_client.get_workflow_run_by_id(workflow_run_id)
-        gathered_context = dict(workflow_run.gathered_context or {})
-        gathered_context["call_id"] = call_id
-        await db_client.update_workflow_run(
-            run_id=workflow_run_id, gathered_context=gathered_context
-        )
-
-    response_content = await provider.get_webhook_response(
-        workflow_id, user_id, workflow_run_id
-    )
-    return HTMLResponse(content=response_content, media_type="application/xml")
-
-
-async def _handle_plivo_status_callback(
-    workflow_run_id: int,
-    request: Request,
-    x_plivo_signature_v3: Optional[str],
-    x_plivo_signature_ma_v3: Optional[str],
-    x_plivo_signature_v3_nonce: Optional[str],
-):
-    set_current_run_id(workflow_run_id)
-
-    form_data = await request.form()
-    callback_data = dict(form_data)
-    logger.info(
-        f"[run {workflow_run_id}] Received Plivo callback: {json.dumps(callback_data)}"
-    )
-
-    workflow_run = await db_client.get_workflow_run_by_id(workflow_run_id)
-    if not workflow_run:
-        logger.warning(f"Workflow run {workflow_run_id} not found for Plivo callback")
-        return {"status": "ignored", "reason": "workflow_run_not_found"}
-
-    workflow = await db_client.get_workflow_by_id(workflow_run.workflow_id)
-    if not workflow:
-        logger.warning(f"Workflow {workflow_run.workflow_id} not found")
-        return {"status": "ignored", "reason": "workflow_not_found"}
-
-    provider = await get_telephony_provider(workflow.organization_id)
-
-    signature = x_plivo_signature_v3 or x_plivo_signature_ma_v3
-    if signature:
-        backend_endpoint, _ = await get_backend_endpoints()
-        callback_kind = request.url.path.split("/")[-2]
-        full_url = f"{backend_endpoint}/api/v1/telephony/plivo/{callback_kind}/{workflow_run_id}"
-        is_valid = await provider.verify_inbound_signature(
-            full_url,
-            callback_data,
-            signature,
-            x_plivo_signature_v3_nonce,
-        )
-        if not is_valid:
-            logger.warning(f"[run {workflow_run_id}] Invalid Plivo webhook signature")
-            return {"status": "error", "reason": "invalid_signature"}
-
-    parsed_data = provider.parse_status_callback(callback_data)
-    status_update = StatusCallbackRequest(
-        call_id=parsed_data["call_id"],
-        status=parsed_data["status"],
-        from_number=parsed_data.get("from_number"),
-        to_number=parsed_data.get("to_number"),
-        direction=parsed_data.get("direction"),
-        duration=parsed_data.get("duration"),
-        extra=parsed_data.get("extra", {}),
-    )
-
-    await _process_status_update(workflow_run_id, status_update)
-    return {"status": "success"}
-
-
-@router.post("/plivo/hangup-callback/{workflow_run_id}")
-async def handle_plivo_hangup_callback(
-    workflow_run_id: int,
-    request: Request,
-    x_plivo_signature_v3: Optional[str] = Header(None),
-    x_plivo_signature_ma_v3: Optional[str] = Header(None),
-    x_plivo_signature_v3_nonce: Optional[str] = Header(None),
-):
-    """Handle Plivo hangup callbacks."""
-    return await _handle_plivo_status_callback(
-        workflow_run_id,
-        request,
-        x_plivo_signature_v3,
-        x_plivo_signature_ma_v3,
-        x_plivo_signature_v3_nonce,
-    )
-
-
-@router.post("/plivo/ring-callback/{workflow_run_id}")
-async def handle_plivo_ring_callback(
-    workflow_run_id: int,
-    request: Request,
-    x_plivo_signature_v3: Optional[str] = Header(None),
-    x_plivo_signature_ma_v3: Optional[str] = Header(None),
-    x_plivo_signature_v3_nonce: Optional[str] = Header(None),
-):
-    """Handle Plivo ring callbacks."""
-    return await _handle_plivo_status_callback(
-        workflow_run_id,
-        request,
-        x_plivo_signature_v3,
-        x_plivo_signature_ma_v3,
-        x_plivo_signature_v3_nonce,
-    )
-
-
-@router.post("/vobiz/hangup-callback/{workflow_run_id}")
-async def handle_vobiz_hangup_callback(
-    workflow_run_id: int,
-    request: Request,
-    x_vobiz_signature: Optional[str] = Header(None),
-    x_vobiz_timestamp: Optional[str] = Header(None),
-):
-    """Handle Vobiz hangup callback (sent when call ends).
-
-    Vobiz sends callbacks to hangup_url when the call terminates.
-    This includes call duration, status, and billing information.
-    """
-    set_current_run_id(workflow_run_id)
-
-    # Logging all headers and body to understand what Vobiz actually sends
-    all_headers = dict(request.headers)
-    logger.info(
-        f"[run {workflow_run_id}] Vobiz hangup callback - Headers: {json.dumps(all_headers)}"
-    )
-
-    # Parse the callback data (Vobiz sends form data or JSON)
-    form_data = await request.form()
-    callback_data = dict(form_data)
-
-    # TODO: Remove this debug logging after Vobiz team clarifies webhook authentication
-    logger.info(
-        f"[run {workflow_run_id}] Vobiz hangup callback - Body: {json.dumps(callback_data)}"
-    )
-    logger.info(
-        f"[run {workflow_run_id}] Received Vobiz hangup callback {json.dumps(callback_data)}"
-    )
-
-    # Verify signature if provided
-    if x_vobiz_signature:
-        # We need the workflow run to get organization for provider credentials
-        workflow_run = await db_client.get_workflow_run_by_id(workflow_run_id)
-        if not workflow_run:
-            logger.warning(
-                f"[run {workflow_run_id}] Workflow run not found for signature verification"
-            )
-            return {"status": "error", "reason": "workflow_run_not_found"}
-
-        workflow = await db_client.get_workflow_by_id(workflow_run.workflow_id)
-        if not workflow:
-            logger.warning(
-                f"[run {workflow_run_id}] Workflow not found for signature verification"
-            )
-            return {"status": "error", "reason": "workflow_not_found"}
-
-        provider = await get_telephony_provider(workflow.organization_id)
-
-        # Get raw body for signature verification
-        raw_body = await request.body()
-        webhook_body = raw_body.decode("utf-8")
-
-        # Verify signature
-        backend_endpoint, _ = await get_backend_endpoints()
-        webhook_url = f"{backend_endpoint}/api/v1/telephony/vobiz/hangup-callback/{workflow_run_id}"
-
-        is_valid = await provider.verify_webhook_signature(
-            webhook_url,
-            callback_data,
-            x_vobiz_signature,
-            x_vobiz_timestamp,
-            webhook_body,
-        )
-
-        if not is_valid:
-            logger.warning(
-                f"[run {workflow_run_id}] Invalid Vobiz hangup callback signature"
-            )
-            return {"status": "error", "reason": "invalid_signature"}
-
-        logger.info(f"[run {workflow_run_id}] Vobiz hangup callback signature verified")
-    else:
-        # Get workflow run for processing (signature verification already got it if needed)
-        workflow_run = await db_client.get_workflow_run_by_id(workflow_run_id)
-    if not workflow_run:
-        logger.warning(
-            f"[run {workflow_run_id}] Workflow run not found for Vobiz hangup callback"
-        )
-        return {"status": "ignored", "reason": "workflow_run_not_found"}
-
-    # Get workflow and provider
-    workflow = await db_client.get_workflow_by_id(workflow_run.workflow_id)
-    if not workflow:
-        logger.warning(f"[run {workflow_run_id}] Workflow not found")
-        return {"status": "ignored", "reason": "workflow_not_found"}
-
-    provider = await get_telephony_provider(workflow.organization_id)
-
-    logger.debug(
-        f"[run {workflow_run_id}] Processing Vobiz hangup with provider: {provider.PROVIDER_NAME}"
-    )
-
-    # Parse the callback data into generic format
-    parsed_data = provider.parse_status_callback(callback_data)
-
-    logger.debug(
-        f"[run {workflow_run_id}] Parsed Vobiz callback data: {json.dumps(parsed_data)}"
-    )
-
-    # Create StatusCallbackRequest from parsed data
-    status_update = StatusCallbackRequest(
-        call_id=parsed_data["call_id"],
-        status=parsed_data["status"],
-        from_number=parsed_data.get("from_number"),
-        to_number=parsed_data.get("to_number"),
-        direction=parsed_data.get("direction"),
-        duration=parsed_data.get("duration"),
-        extra=parsed_data.get("extra", {}),
-    )
-
-    # Process the status update
-    await _process_status_update(workflow_run_id, status_update)
-
-    logger.info(f"[run {workflow_run_id}] Vobiz hangup callback processed successfully")
-
-    return {"status": "success"}
-
-
-@router.post("/vobiz/ring-callback/{workflow_run_id}")
-async def handle_vobiz_ring_callback(
-    workflow_run_id: int,
-    request: Request,
-    x_vobiz_signature: Optional[str] = Header(None),
-    x_vobiz_timestamp: Optional[str] = Header(None),
-):
-    """Handle Vobiz ring callback (sent when call starts ringing).
-
-    Vobiz can send callbacks to ring_url when the call starts ringing.
-    This is optional and used for tracking ringing status.
-    """
-    set_current_run_id(workflow_run_id)
-
-    # Logging all headers and body to understand what Vobiz actually sends
-    all_headers = dict(request.headers)
-    logger.info(
-        f"[run {workflow_run_id}] Vobiz ring callback - Headers: {json.dumps(all_headers)}"
-    )
-
-    # Parse the callback data
-    form_data = await request.form()
-    callback_data = dict(form_data)
-
-    # TODO: Remove this debug logging after Vobiz team clarifies webhook authentication
-    logger.info(
-        f"[run {workflow_run_id}] Vobiz ring callback - Body: {json.dumps(callback_data)}"
-    )
-
-    logger.info(
-        f"[run {workflow_run_id}] Received Vobiz ring callback {json.dumps(callback_data)}"
-    )
-
-    # Verify signature if provided
-    if x_vobiz_signature:
-        # We need the workflow run to get organization for provider credentials
-        workflow_run = await db_client.get_workflow_run_by_id(workflow_run_id)
-        if not workflow_run:
-            logger.warning(
-                f"[run {workflow_run_id}] Workflow run not found for signature verification"
-            )
-            return {"status": "error", "reason": "workflow_run_not_found"}
-
-        workflow = await db_client.get_workflow_by_id(workflow_run.workflow_id)
-        if not workflow:
-            logger.warning(
-                f"[run {workflow_run_id}] Workflow not found for signature verification"
-            )
-            return {"status": "error", "reason": "workflow_not_found"}
-
-        provider = await get_telephony_provider(workflow.organization_id)
-
-        # Get raw body for signature verification
-        raw_body = await request.body()
-        webhook_body = raw_body.decode("utf-8")
-
-        # Verify signature
-        backend_endpoint, _ = await get_backend_endpoints()
-        webhook_url = (
-            f"{backend_endpoint}/api/v1/telephony/vobiz/ring-callback/{workflow_run_id}"
-        )
-
-        is_valid = await provider.verify_webhook_signature(
-            webhook_url,
-            callback_data,
-            x_vobiz_signature,
-            x_vobiz_timestamp,
-            webhook_body,
-        )
-
-        if not is_valid:
-            logger.warning(
-                f"[run {workflow_run_id}] Invalid Vobiz ring callback signature"
-            )
-            return {"status": "error", "reason": "invalid_signature"}
-
-        logger.info(f"[run {workflow_run_id}] Vobiz ring callback signature verified")
-    else:
-        # Get workflow run for processing (signature verification already got it if needed)
-        workflow_run = await db_client.get_workflow_run_by_id(workflow_run_id)
-    if not workflow_run:
-        logger.warning(
-            f"[run {workflow_run_id}] Workflow run not found for Vobiz ring callback"
-        )
-        return {"status": "ignored", "reason": "workflow_run_not_found"}
-
-    # Log the ringing event
-    telephony_callback_logs = workflow_run.logs.get("telephony_status_callbacks", [])
-    ring_log = {
-        "status": "ringing",
-        "timestamp": datetime.now(UTC).isoformat(),
-        "call_id": callback_data.get("call_uuid", callback_data.get("CallUUID", "")),
-        "event_type": "ring",
-        "raw_data": callback_data,
-    }
-    telephony_callback_logs.append(ring_log)
-
-    # Update workflow run logs
-    await db_client.update_workflow_run(
-        run_id=workflow_run_id,
-        logs={"telephony_status_callbacks": telephony_callback_logs},
-    )
-
-    logger.info(f"[run {workflow_run_id}] Vobiz ring callback logged")
-
-    return {"status": "success"}
-
-
-@router.post("/cloudonix/status-callback/{workflow_run_id}")
-async def handle_cloudonix_status_callback(
-    workflow_run_id: int,
-    request: Request,
-):
-    """Handle Cloudonix-specific status callbacks.
-
-    Cloudonix sends call status updates to the callback URL specified during call initiation.
-    """
-    set_current_run_id(workflow_run_id)
-    # Parse callback data - determine if JSON or form data
-    content_type = request.headers.get("content-type", "")
-
-    if "application/json" in content_type:
-        callback_data = await request.json()
-    else:
-        # Assume form data (like Twilio)
-        form_data = await request.form()
-        callback_data = dict(form_data)
-
-    logger.info(
-        f"[run {workflow_run_id}] Received Cloudonix status callback: {json.dumps(callback_data)}"
-    )
-
-    # Get workflow run to find organization
-    workflow_run = await db_client.get_workflow_run_by_id(workflow_run_id)
-    if not workflow_run:
-        logger.warning(f"Workflow run {workflow_run_id} not found for status callback")
-        return {"status": "ignored", "reason": "workflow_run_not_found"}
-
-    # Get workflow and provider
-    workflow = await db_client.get_workflow_by_id(workflow_run.workflow_id)
-    if not workflow:
-        logger.warning(f"Workflow {workflow_run.workflow_id} not found")
-        return {"status": "ignored", "reason": "workflow_not_found"}
-
-    provider = await get_telephony_provider(workflow.organization_id)
-
-    # Parse the callback data into generic format
-    parsed_data = provider.parse_status_callback(callback_data)
-
-    # Create StatusCallbackRequest from parsed data
-    status_update = StatusCallbackRequest(
-        call_id=parsed_data["call_id"],
-        status=parsed_data["status"],
-        from_number=parsed_data.get("from_number"),
-        to_number=parsed_data.get("to_number"),
-        direction=parsed_data.get("direction"),
-        duration=parsed_data.get("duration"),
-        extra=parsed_data.get("extra", {}),
-    )
-
-    # Process the status update
-    await _process_status_update(workflow_run_id, status_update)
-
-    return {"status": "success"}
-
-
-@router.post("/vobiz/hangup-callback/workflow/{workflow_id}")
-async def handle_vobiz_hangup_callback_by_workflow(
-    workflow_id: int,
-    request: Request,
-    x_vobiz_signature: Optional[str] = Header(None),
-    x_vobiz_timestamp: Optional[str] = Header(None),
-):
-    """Handle Vobiz hangup callback with workflow_id - finds workflow run by call_id."""
-
-    all_headers = dict(request.headers)
-    logger.info(
-        f"[workflow {workflow_id}] Vobiz hangup callback - Headers: {json.dumps(all_headers)}"
-    )
-
-    try:
-        callback_data, _ = await parse_webhook_request(request)
-    except ValueError:
-        callback_data = {}
-
-    call_uuid = callback_data.get("CallUUID") or callback_data.get("call_uuid")
-    logger.info(
-        f"[workflow {workflow_id}] Received Vobiz hangup callback for call {call_uuid}: {json.dumps(callback_data)}"
-    )
-
-    if not call_uuid:
-        logger.warning(
-            f"[workflow {workflow_id}] No call_uuid found in Vobiz hangup callback"
-        )
-        return {"status": "error", "message": "No call_uuid found"}
-
-    workflow_client = WorkflowClient()
-    workflow = await workflow_client.get_workflow_by_id(workflow_id)
-    if not workflow:
-        logger.warning(f"[workflow {workflow_id}] Workflow not found")
-        return {"status": "error", "message": "workflow_not_found"}
-
-    provider = await get_telephony_provider(workflow.organization_id)
-
-    if x_vobiz_signature:
-        raw_body = await request.body()
-        webhook_body = raw_body.decode("utf-8")
-        backend_endpoint, _ = await get_backend_endpoints()
-        webhook_url = f"{backend_endpoint}/api/v1/telephony/vobiz/hangup-callback/workflow/{workflow_id}"
-
-        is_valid = await provider.verify_webhook_signature(
-            webhook_url,
-            callback_data,
-            x_vobiz_signature,
-            x_vobiz_timestamp,
-            webhook_body,
-        )
-
-        if not is_valid:
-            logger.warning(
-                f"[workflow {workflow_id}] Invalid Vobiz hangup callback signature"
-            )
-            return {"status": "error", "message": "invalid_signature"}
-
-        logger.info(
-            f"[workflow {workflow_id}] Vobiz hangup callback signature verified"
-        )
-
-    try:
-        db_client = WorkflowRunClient()
-        async with db_client.async_session() as session:
-            # Fetch workflow run with matching call_id in gathered_context
-            query = text("""
-                SELECT id FROM workflow_runs 
-                WHERE workflow_id = :workflow_id 
-                AND CAST(gathered_context AS jsonb) @> CAST(:call_id_json AS jsonb)
-                ORDER BY created_at DESC 
-                LIMIT 1
-            """)
-
-            result = await session.execute(
-                query,
-                {
-                    "workflow_id": workflow_id,
-                    "call_id_json": json.dumps({"call_id": call_uuid}),
-                },
-            )
-            workflow_run_row = result.fetchone()
-
-            if not workflow_run_row:
-                logger.warning(
-                    f"[workflow {workflow_id}] No workflow run found for call {call_uuid}"
-                )
-                return {"status": "ignored", "reason": "workflow_run_not_found"}
-
-            workflow_run_id = workflow_run_row[0]
-            set_current_run_id(workflow_run_id)
-            logger.info(
-                f"[workflow {workflow_id}] Found workflow run {workflow_run_id} for call {call_uuid}"
-            )
-
-    except Exception as e:
-        logger.error(
-            f"[workflow {workflow_id}] Error finding workflow run for call {call_uuid}: {e}"
-        )
-        return {"status": "error", "message": str(e)}
-
-    try:
-        workflow_run = await db_client.get_workflow_run_by_id(workflow_run_id)
-        if not workflow_run:
-            logger.warning(f"[run {workflow_run_id}] Workflow run not found")
-            return {"status": "ignored", "reason": "workflow_run_not_found"}
-
-        parsed_data = provider.parse_status_callback(callback_data)
-
-        status = StatusCallbackRequest(
-            call_id=parsed_data["call_id"],
-            status=parsed_data["status"],
-            from_number=parsed_data.get("from_number"),
-            to_number=parsed_data.get("to_number"),
-            direction=parsed_data.get("direction"),
-            duration=parsed_data.get("duration"),
-            extra=parsed_data.get("extra", {}),
-        )
-
-        await _process_status_update(workflow_run_id, status)
-
-        logger.info(
-            f"[run {workflow_run_id}] Vobiz hangup callback processed successfully"
-        )
-        return {"status": "success"}
-
-    except Exception as e:
-        logger.error(
-            f"[run {workflow_run_id}] Error processing Vobiz hangup callback: {e}"
-        )
-        return {"status": "error", "message": str(e)}
 
 
 @router.post("/inbound/{workflow_id}")
@@ -1601,59 +761,6 @@ async def handle_inbound_fallback(request: Request):
         )
 
         return generic_hangup_response()
-
-
-@router.post("/cloudonix/cdr")
-async def handle_cloudonix_cdr(request: Request):
-    """Handle Cloudonix CDR (Call Detail Record) webhooks.
-
-    Cloudonix sends CDR records when calls complete. The CDR contains:
-    - domain: Used to identify the organization
-    - call_id: Used to find the workflow run
-    - disposition: Call termination status (ANSWER, BUSY, CANCEL, FAILED, CONGESTION, NOANSWER)
-    - duration/billsec: Call duration information
-    """
-    try:
-        cdr_data = await request.json()
-    except Exception as e:
-        logger.error(f"Failed to parse Cloudonix CDR JSON: {e}")
-        return {"status": "error", "message": "Invalid JSON payload"}
-
-    # Extract domain to find organization
-    domain = cdr_data.get("domain")
-    if not domain:
-        logger.warning("Cloudonix CDR missing domain field")
-        return {"status": "error", "message": "Missing domain field"}
-
-    # Extract call_id to find workflow run
-    call_id = cdr_data.get("session").get("token")
-    logger.info(f"Cloudonix CDR data for call id {call_id} - {cdr_data}")
-    if not call_id:
-        logger.warning("Cloudonix CDR missing call_id field")
-        return {"status": "error", "message": "Missing call_id field"}
-
-    # Find workflow run by call_id in gathered_context
-    workflow_run = await db_client.get_workflow_run_by_call_id(call_id)
-    if not workflow_run:
-        logger.warning(f"No workflow run found for Cloudonix call_id: {call_id}")
-        return {"status": "ignored", "reason": "workflow_run_not_found"}
-
-    workflow_run_id = workflow_run.id
-    set_current_run_id(workflow_run_id)
-    logger.info(f"[run {workflow_run_id}] Processing Cloudonix CDR for call {call_id}")
-
-    # Convert CDR to status update using StatusCallbackRequest
-    status_update = StatusCallbackRequest.from_cloudonix_cdr(cdr_data)
-
-    # Process the status update
-    await _process_status_update(workflow_run_id, status_update)
-
-    logger.info(
-        f"[run {workflow_run_id}] Cloudonix CDR processed successfully - "
-        f"disposition: {cdr_data.get('disposition')}, status: {status_update.status}"
-    )
-
-    return {"status": "success"}
 
 
 class TransferCallRequest(BaseModel):
