+++ api/services/telephony/factory.py
@@ -11,11 +11,25 @@ from loguru import logger
 
 from api.db import db_client
 from api.enums import OrganizationConfigurationKey
-from api.services.telephony import (
-    providers as _providers,  # noqa: F401  -- triggers registration
-)
 from api.services.telephony import registry
 from api.services.telephony.base import TelephonyProvider
+
+_providers_loaded = False
+
+
+def _ensure_providers_loaded() -> None:
+    """Lazy-import the providers package to trigger registration.
+
+    Importing at module load time would create a cycle: provider packages
+    import their own ``routes.py``, which imports ``get_telephony_provider``
+    from this module — partially-initialized when the providers package
+    runs. Deferring until the first factory call breaks the cycle without
+    making provider authors think about lazy imports.
+    """
+    global _providers_loaded
+    if not _providers_loaded:
+        from api.services.telephony import providers as _  # noqa: F401
+        _providers_loaded = True
 
 
 async def load_telephony_config(organization_id: int) -> Dict[str, Any]:
@@ -25,6 +39,7 @@ async def load_telephony_config(organization_id: int) -> Dict[str, Any]:
     keys plus a ``provider`` discriminator). Raises ValueError if no config
     is stored or the provider is unknown.
     """
+    _ensure_providers_loaded()
     if not organization_id:
         raise ValueError("Organization ID is required to load telephony configuration")
 
@@ -56,4 +71,5 @@ async def get_telephony_provider(organization_id: int) -> TelephonyProvider:
 
 async def get_all_telephony_providers() -> List[Type[TelephonyProvider]]:
     """Return all registered telephony provider classes for webhook detection."""
+    _ensure_providers_loaded()
     return [spec.provider_cls for spec in registry.all_specs()]
