+++ api/services/telephony/providers/cloudonix/__init__.py
@@ -7,6 +7,7 @@ from api.services.telephony.registry import ProviderSpec, register
 
 from .config import CloudonixConfigurationRequest, CloudonixConfigurationResponse
 from .provider import CloudonixProvider
+from .routes import router as routes_router
 from .transport import create_transport
 
 
@@ -36,6 +37,7 @@ SPEC = ProviderSpec(
     transport_factory=create_transport,
     audio_config=_AUDIO_CONFIG,
     config_request_cls=CloudonixConfigurationRequest,
+    router=routes_router,
     config_response_cls=CloudonixConfigurationResponse,
 )
