+++ api/services/telephony/providers/twilio/routes.py
@@ -1,0 +1,112 @@
+"""Twilio telephony routes (webhooks, status callbacks, answer URLs).
+
+Mounted under ``/api/v1/telephony`` by ``api.routes.telephony`` via the
+provider registry — see ProviderSpec.router.
+"""
+
+import json
+from datetime import UTC, datetime
+from typing import Optional
+
+from fastapi import APIRouter, Header, HTTPException, Request, Response
+from loguru import logger
+from starlette.responses import HTMLResponse
+
+from api.db import db_client
+from api.errors.telephony_errors import TelephonyError
+from api.services.telephony.factory import get_telephony_provider
+from api.services.telephony.status_processor import (
+    StatusCallbackRequest,
+    _process_status_update,
+)
+from api.utils.common import get_backend_endpoints
+from api.utils.telephony_helper import (
+    generic_hangup_response,
+    normalize_webhook_data,
+    parse_webhook_request,
+)
+from pipecat.utils.run_context import set_current_run_id
+
+router = APIRouter()
+
+
+@router.post("/twiml", include_in_schema=False)
+async def handle_twiml_webhook(
+    workflow_id: int, user_id: int, workflow_run_id: int, organization_id: int
+):
+    """
+    Handle initial webhook from telephony provider.
+    Returns provider-specific response (e.g., TwiML for Twilio).
+    """
+
+    provider = await get_telephony_provider(organization_id)
+
+    response_content = await provider.get_webhook_response(
+        workflow_id, user_id, workflow_run_id
+    )
+
+    return HTMLResponse(content=response_content, media_type="application/xml")
+
+@router.post("/twilio/status-callback/{workflow_run_id}")
+async def handle_twilio_status_callback(
+    workflow_run_id: int,
+    request: Request,
+    x_webhook_signature: Optional[str] = Header(None),
+):
+    """Handle Twilio-specific status callbacks."""
+    set_current_run_id(workflow_run_id)
+
+    # Parse form data
+    form_data = await request.form()
+    callback_data = dict(form_data)
+
+    logger.info(
+        f"[run {workflow_run_id}] Received status callback: {json.dumps(callback_data)}"
+    )
+
+    # Get workflow run to find organization
+    workflow_run = await db_client.get_workflow_run_by_id(workflow_run_id)
+    if not workflow_run:
+        logger.warning(f"Workflow run {workflow_run_id} not found for status callback")
+        return {"status": "ignored", "reason": "workflow_run_not_found"}
+
+    # Get workflow and provider
+    workflow = await db_client.get_workflow_by_id(workflow_run.workflow_id)
+    if not workflow:
+        logger.warning(f"Workflow {workflow_run.workflow_id} not found")
+        return {"status": "ignored", "reason": "workflow_not_found"}
+
+    provider = await get_telephony_provider(workflow.organization_id)
+
+    if x_webhook_signature:
+        backend_endpoint, _ = await get_backend_endpoints()
+        full_url = f"{backend_endpoint}/api/v1/telephony/twilio/status-callback/{workflow_run_id}"
+
+        is_valid = await provider.verify_webhook_signature(
+            full_url, callback_data, x_webhook_signature
+        )
+
+        if not is_valid:
+            logger.warning(
+                f"Invalid webhook signature for workflow run {workflow_run_id}"
+            )
+            return {"status": "error", "reason": "invalid_signature"}
+
+    # Parse the callback data into generic format
+    parsed_data = provider.parse_status_callback(callback_data)
+
+    # Create StatusCallbackRequest from parsed data
+    status_update = StatusCallbackRequest(
+        call_id=parsed_data["call_id"],
+        status=parsed_data["status"],
+        from_number=parsed_data.get("from_number"),
+        to_number=parsed_data.get("to_number"),
+        direction=parsed_data.get("direction"),
+        duration=parsed_data.get("duration"),
+        extra=parsed_data.get("extra", {}),
+    )
+
+    # Process the status update
+    await _process_status_update(workflow_run_id, status_update)
+
+    return {"status": "success"}
