+++ api/services/telephony/providers/vonage/__init__.py
@@ -7,6 +7,7 @@ from api.services.telephony.registry import ProviderSpec, register
 
 from .config import VonageConfigurationRequest, VonageConfigurationResponse
 from .provider import VonageProvider
+from .routes import router as routes_router
 from .transport import create_transport
 
 
@@ -37,6 +38,7 @@ SPEC = ProviderSpec(
     transport_factory=create_transport,
     audio_config=_AUDIO_CONFIG,
     config_request_cls=VonageConfigurationRequest,
+    router=routes_router,
     config_response_cls=VonageConfigurationResponse,
 )
